change upload firmware to install from local directory by default

+++ SMBR.Utilities/reactor-firmware-update/update_firmware.sh
@@ -2,6 +2,7 @@
 
 usage() {
     echo "Usage: $0 [-l local_firmware_path] [-f]"
+    echo "  Without arguments it uses existing firmware from firmware directory"
     echo "  -l : Path to local firmware zip file"
     echo "  -f : Fetch latest firmware from GitHub"
     echo "  -h : Show this help"
@@ -13,6 +14,7 @@ firmware_dir="./firmware"
 
 # Parse command line arguments
 local_firmware=""
+install_existing=false
 fetch_firmware=false
 while getopts "l:fh" opt; do
     case $opt in
@@ -25,39 +27,42 @@ done
 
 # Ensure at least one option is specified
 if [ -z "$local_firmware" ] && [ "$fetch_firmware" = false ]; then
-    echo "Warning: No action specified"
-    usage
+    install_existing=true
+    echo "Install existing firmware"
 fi
 
 # Ensure the firmware directory exists
 mkdir -p "$firmware_dir"
 
-if [ -n "$local_firmware" ]; then
-    if [ ! -f "$local_firmware" ]; then
-        echo "Error: Local firmware file not found: $local_firmware"
-        exit 1
-    fi
-    echo "Using local firmware: $local_firmware"
-    cp "$local_firmware" "$firmware_dir/$(basename "$local_firmware")"
-    archive_name="$(basename "$local_firmware")"
-elif [ "$fetch_firmware" = true ]; then
-    # Fetch the latest release URL
-    url=$(curl -s https://api.github.com/repos/TrendBit/SMBR-firmware/releases/latest | grep "browser_download_url.*\.zip" | cut -d '"' -f 4)
 
-    if [ -n "$url" ]; then
-        echo "Downloading from: $url"
-        # Download the archive
-        archive_name="$(basename "$url")"
-        curl -L -o "$firmware_dir/$archive_name" "$url"
-    else
-        echo "Error: Could not fetch the download URL."
-        exit 1
+if [ "$install_existing" = false ]; then
+    if [ -n "$local_firmware" ]; then
+        if [ ! -f "$local_firmware" ]; then
+            echo "Error: Local firmware file not found: $local_firmware"
+            exit 1
+        fi
+        echo "Using local firmware: $local_firmware"
+        cp "$local_firmware" "$firmware_dir/$(basename "$local_firmware")"
+        archive_name="$(basename "$local_firmware")"
+    elif [ "$fetch_firmware" = true ]; then
+        # Fetch the latest release URL
+        url=$(curl -s https://api.github.com/repos/TrendBit/SMBR-firmware/releases/latest | grep "browser_download_url.*\.zip" | cut -d '"' -f 4)
+
+        if [ -n "$url" ]; then
+            echo "Downloading from: $url"
+            # Download the archive
+            archive_name="$(basename "$url")"
+            curl -L -o "$firmware_dir/$archive_name" "$url"
+        else
+            echo "Error: Could not fetch the download URL."
+            exit 1
+        fi
     fi
+
+    # Unzip the archive into the firmware directory
+    unzip -o "$firmware_dir/$archive_name" -d "$firmware_dir"
+    echo "Firmware unzipped to $firmware_dir"
 fi
-
-# Unzip the archive into the firmware directory
-unzip -o "$firmware_dir/$archive_name" -d "$firmware_dir"
-echo "Firmware unzipped to $firmware_dir"
 
 # Stop the services
 echo "Stopping services..."
